Data-transformation(updated)

+++ model/Img-subset.py
@@ -23,27 +23,36 @@ if not os.path.exists(Val_dir):
 if not os.path.exists(Val_dirl):
     os.mkdir(Val_dirl)
 files = glob.glob(os.path.join(image_dir, '*.jpg'))
-for fil in range(0,25001,5):
-    basename = os.path.basename(files[fil])
-    filename = os.path.splitext(basename)[0]
-    if  os.path.exists(os.path.join(input_dir, f"{filename}.txt")):
-        shutil.copyfile(os.path.join(image_dir,basename), os.path.join(Train_dir,basename))
-        shutil.copyfile(os.path.join(input_dir,f"{filename}.txt"), os.path.join(Train_dirl,f"{filename}.txt"))
-        os.remove(os.path.join(image_dir,basename))
-        os.remove(os.path.join(input_dir,f"{filename}.txt"))
-for fill in range(-1,10001,5):
-    basename = os.path.basename(files[fill])
-    filename = os.path.splitext(basename)[0]
-    if  os.path.exists(os.path.join(input_dir, f"{filename}.txt")):
-        shutil.copyfile(os.path.join(image_dir,basename), f"{os.path.join(Val_dir,basename)}")
-        shutil.copyfile(os.path.join(input_dir,f"{filename}.txt"), os.path.join(Val_dirl,f"{filename}.txt"))
-        os.remove(os.path.join(image_dir,basename))
-        os.remove(os.path.join(input_dir,f"{filename}.txt"))
-for fill in range(0,5001,5):
-    basename = os.path.basename(files[fill])
-    filename = os.path.splitext(basename)[0]
-    if  os.path.exists(os.path.join(input_dir, f"{filename}.txt")):
-        shutil.copyfile(os.path.join(image_dir,basename), f"{os.path.join(test_dir,basename)}")
-        shutil.copyfile(os.path.join(input_dir,f"{filename}.txt"), os.path.join(test_dirl,f"{filename}.txt"))
-        os.remove(os.path.join(image_dir,basename))
-        os.remove(os.path.join(input_dir,f"{filename}.txt"))
+filess = glob.glob(os.path.join(Train_dir, '*.jpg'))
+filesss = glob.glob(os.path.join(Val_dir, '*.jpg'))
+filessss = glob.glob(os.path.join(test_dir, '*.jpg'))
+cTr=filess.count()
+cVal=filesss.count()
+cTest=filessss.count()
+if cTr!= range(0,25001,5):
+    for fil in range(0,25001-(cTr*5),5):
+        basename = os.path.basename(files[fil])
+        filename = os.path.splitext(basename)[0]
+        if  os.path.exists(os.path.join(input_dir, f"{filename}.txt")):
+            shutil.copyfile(os.path.join(image_dir,basename), os.path.join(Train_dir,basename))
+            shutil.copyfile(os.path.join(input_dir,f"{filename}.txt"), os.path.join(Train_dirl,f"{filename}.txt"))
+            os.remove(os.path.join(image_dir,basename))
+            os.remove(os.path.join(input_dir,f"{filename}.txt"))
+if cVal!= range(0,20001,5):
+    for fill in range(-1,10001-(cVal*5),5):
+        basename = os.path.basename(files[fill])
+        filename = os.path.splitext(basename)[0]
+        if  os.path.exists(os.path.join(input_dir, f"{filename}.txt")):
+            shutil.copyfile(os.path.join(image_dir,basename), os.path.join(Val_dir,basename))
+            shutil.copyfile(os.path.join(input_dir,f"{filename}.txt"), os.path.join(Val_dirl,f"{filename}.txt"))
+            os.remove(os.path.join(image_dir,basename))
+            os.remove(os.path.join(input_dir,f"{filename}.txt"))
+if cTest!= range(0,5001,5):
+    for fill in range(0,5001-(cTest*5),5):
+        basename = os.path.basename(files[fill])
+        filename = os.path.splitext(basename)[0]
+        if  os.path.exists(os.path.join(input_dir, f"{filename}.txt")):
+            shutil.copyfile(os.path.join(image_dir,basename), os.path.join(test_dir,basename))
+            shutil.copyfile(os.path.join(input_dir,f"{filename}.txt"), os.path.join(test_dirl,f"{filename}.txt"))
+            os.remove(os.path.join(image_dir,basename))
+            os.remove(os.path.join(input_dir,f"{filename}.txt"))
